chore: re-expose interface and error in dbkit pkg
