+++ Taskfile.yml
@@ -29,3 +29,16 @@ tasks:
     desc: setup test db
     cmds:
       - ./scripts/testenv.sh
+
+  mock:
+    desc: Auto-generate mocks for all interfaces
+    cmds:
+      - rm -rf test/mocks/* # 💥 Clean old mocks
+      - mockery
+      - go mod tidy
+    silent: false
+
+  deps:
+    desc: install dependencies
+    cmds:
+      - go install github.com/vektra/mockery/v3@v3.4.0
